+++ doc/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 * (alerts) Added a threshold for change in voting power to reduce alert spam in the case of tiny changes. The default is **1** and this can be customised by modifying the `change_in_voting_power_threshold` field in the `internal_config.ini`.
 * (errors) Improved handling of IncompleteRead errors that were showing up on the alert channels.
+* (alerts/peers) Safe and danger boundary can now be `-1` and `-2` respectively, which causes peer alerts to be disabled since `peers > safe > danger` for any number of peers (`>= 0`).
 
 ## 1.1.2